Updated project ID for "BankToBeTested"
Updated the project reference ID for "BankToBeTested" in the "UnitTestProject1.csproj" file and the project ID in the "UnitTestProject1.sln" file. Also, the project configuration platforms for "BankToBeTested" in the "UnitTestProject1.sln" file have been updated to reflect the new project ID, and the configurations for the old project ID have been removed.