Added `DepositBalanceZero` test in `Accounts` class
Added a new test class `Accounts` in the `UnitTestProject1` namespace. Inside this class, a new test method `DepositBalanceZero` was created. This method creates an `Account` object with an id of 1, sets a `value` variable to 0.0, and performs a transaction on the `account` object with the `value` variable. Two assertions are made to check if the transaction was successful and if the balance of the account is 0.0 after the transaction.