`Modify methods and conditions in Program.cs and UnitTest.cs`
Modified several methods and conditions in `Program.cs` and `UnitTest.cs` under the `UnitTestProject1` namespace. In `Program.cs`, wrapped the initialization of `accounts[i]` in a block, commented out `ListAccounts()` and `FindAccount(int id)` methods, and changed the `FindAccount(int id)` method to use a `foreach` loop. Also, commented the condition in the `AddTransaction()` method and added an `else if` condition in the `Transaction()` method to check if `balance` is greater than or equal to `val`. The constructor of the `Transaction` class has also been commented out. In `UnitTest.cs`, changed the `toAccount` in the `TestAddTransaction_InsufficientBalance()` method and an unnamed method to find an account with a different id.